Fix logging init in lint.py and find_latest_image.py (#7755)
Move init_logging() to separate file and import it into all the CLIs.
