+++ testsuite/lint.py
@@ -1,19 +1,29 @@
-from forge import LocalShell
+# Helm linter.
 
+import logging
 import re
 from typing import Tuple
 from pathlib import Path
 
 import click
 
+from forge import LocalShell
+from applogging import logger, init_logging
+
 
 @click.group()
-def main() -> None:
-    pass
+@click.option(
+    "--log-metadata/--no-log-metadata",
+    default=True,
+)
+@logger
+def main(log_metadata: bool) -> None:
+    init_logging(logger=log, level=logging.DEBUG, print_metadata=log_metadata)
 
 
 @main.command()
 @click.argument("paths", nargs=-1)
+@logger
 def helm(paths: Tuple[str]) -> None:
     shell = LocalShell()
 
@@ -28,7 +38,7 @@ def helm(paths: Tuple[str]) -> None:
                 )
                 if match:
                     fullpath = Path(path).parent / match.group("filename")
-                    print(
+                    log.error(
                         "::error file={fullpath},line={line},col=1::{message}".format(
                             fullpath=fullpath, **match.groupdict()
                         )
